Refactor NSGA-II runner class
