+++ jmetal/algorithm/multiobjective/randomSearch.py
@@ -1,6 +1,6 @@
 from typing import TypeVar, List, Generic
 
-from jmetal.component.nondominatedsolutionlistarchive import NonDominatedSolutionListArchive
+from jmetal.component.archive import NonDominatedSolutionListArchive
 from jmetal.core.problem import Problem
 
 """ Class representing simple random search algorithms """
