+++ jmetal/runner/multiobjective/nsgaIII_standard_settings.py
@@ -15,16 +15,12 @@ from jmetal.util.solution_list_output import SolutionListOutput
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
-S = TypeVar('S')
-R = TypeVar(List[S])
-
 
 def main():
-    float_example()
-    float_example_changing_the_stopping_condition()
+    nsgaii_runner()
 
 
-def float_example() -> None:
+def nsgaii_runner() -> None:
     problem = Kursawe()
     algorithm = NSGAII[FloatSolution, List[FloatSolution]](
         problem,
@@ -36,37 +32,13 @@ def float_example() -> None:
 
     algorithm.run()
     result = algorithm.get_result()
-    logger.info("Algorithm (continuous problem): " + algorithm.get_name())
-    logger.info("Problem: " + problem.get_name())
-    logger.info("Computing time: " + str(algorithm.total_computing_time))
 
     SolutionListOutput[FloatSolution].print_function_values_to_screen(result)
     SolutionListOutput[FloatSolution].print_function_values_to_file("FUN."+problem.get_name(), result)
 
-
-def float_example_changing_the_stopping_condition() -> None:
-    problem = Fonseca()
-
-    class NSGA2b(NSGAII[S, R]):
-        def is_stopping_condition_reached(self):
-            return self.get_current_computing_time() > 4
-
-    algorithm = NSGA2b[FloatSolution, List[FloatSolution]](
-        problem,
-        population_size = 100,
-        max_evaluations = 25000,
-        mutation = Polynomial(1.0/problem.number_of_variables, distribution_index=20),
-        crossover = SBX(1.0, distribution_index=20),
-        selection = BinaryTournament())
-
-    algorithm.run()
-    result = algorithm.get_result()
     logger.info("Algorithm (continuous problem): " + algorithm.get_name())
     logger.info("Problem: " + problem.get_name())
     logger.info("Computing time: " + str(algorithm.total_computing_time))
-
-    SolutionListOutput[FloatSolution].print_function_values_to_screen(result)
-    SolutionListOutput[FloatSolution].print_function_values_to_file("FUN."+problem.get_name(), result)
 
 if __name__ == '__main__':
     main()
